+++ src/lib.rs
@@ -61,6 +61,7 @@
 //! }
 //! ```
 
+use bitflags::bitflags;
 use std::borrow::Borrow;
 use std::convert::TryFrom;
 use std::ffi::{self, CStr, CString};
@@ -211,8 +212,102 @@ impl From<std::io::ErrorKind> for Error {
     }
 }
 
+bitflags! {
+    /// Network device flags.
+    pub struct IfFlags: u32 {
+        /// Set if the device is a loopback interface
+        const LOOPBACK = raw::PCAP_IF_LOOPBACK;
+        /// Set if the device is up
+        const UP = raw::PCAP_IF_UP;
+        /// Set if the device is running
+        const RUNNING = raw::PCAP_IF_RUNNING;
+        /// Set if the device is a wireless interface; this includes IrDA as well as radio-based
+        /// networks such as IEEE 802.15.4 and IEEE 802.11, so it doesn't just mean Wi-Fi
+        const WIRELESS = raw::PCAP_IF_WIRELESS;
+    }
+}
+
+impl From<u32> for IfFlags {
+    fn from(flags: u32) -> Self {
+        IfFlags::from_bits_truncate(flags)
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+/// Indication of whether the adapter is connected or not; for wireless interfaces, "connected"
+/// means "associated with a network".
+pub enum ConnectionStatus {
+    /// It's unknown whether the adapter is connected or not
+    Unknown,
+    /// The adapter is connected
+    Connected,
+    /// The adapter is disconnected
+    Disconnected,
+    /// The notion of "connected" and "disconnected" don't apply to this interface; for example, it
+    /// doesn't apply to a loopback device
+    NotApplicable,
+}
+
+impl From<u32> for ConnectionStatus {
+    fn from(flags: u32) -> Self {
+        match flags & raw::PCAP_IF_CONNECTION_STATUS {
+            raw::PCAP_IF_CONNECTION_STATUS_UNKNOWN => ConnectionStatus::Unknown,
+            raw::PCAP_IF_CONNECTION_STATUS_CONNECTED => ConnectionStatus::Connected,
+            raw::PCAP_IF_CONNECTION_STATUS_DISCONNECTED => ConnectionStatus::Disconnected,
+            raw::PCAP_IF_CONNECTION_STATUS_NOT_APPLICABLE => ConnectionStatus::NotApplicable,
+            // DeviceFlags::CONNECTION_STATUS should be a 2-bit mask which means that the four
+            // values should cover all the possibilities.
+            _ => unreachable!(),
+        }
+    }
+}
+
 #[derive(Debug, Clone)]
-/// A network device name and (potentially) pcap's description of it.
+pub struct DeviceFlags {
+    pub if_flags: IfFlags,
+    pub connection_status: ConnectionStatus,
+}
+
+impl From<u32> for DeviceFlags {
+    fn from(flags: u32) -> Self {
+        DeviceFlags {
+            if_flags: flags.into(),
+            connection_status: flags.into(),
+        }
+    }
+}
+
+impl DeviceFlags {
+    pub fn empty() -> Self {
+        DeviceFlags {
+            if_flags: IfFlags::empty(),
+            connection_status: ConnectionStatus::Unknown,
+        }
+    }
+
+    pub fn contains(&self, if_flags: IfFlags) -> bool {
+        self.if_flags.contains(if_flags)
+    }
+
+    pub fn is_loopback(&self) -> bool {
+        self.contains(IfFlags::LOOPBACK)
+    }
+
+    pub fn is_up(&self) -> bool {
+        self.contains(IfFlags::UP)
+    }
+
+    pub fn is_running(&self) -> bool {
+        self.contains(IfFlags::RUNNING)
+    }
+
+    pub fn is_wireless(&self) -> bool {
+        self.contains(IfFlags::WIRELESS)
+    }
+}
+
+#[derive(Debug, Clone)]
+/// A network device name and pcap's description of it.
 pub struct Device {
     /// The name of the interface
     pub name: String,
@@ -220,14 +315,22 @@ pub struct Device {
     pub desc: Option<String>,
     /// Addresses associated with this interface
     pub addresses: Vec<Address>,
+    /// Interface flags
+    pub flags: DeviceFlags,
 }
 
 impl Device {
-    fn new(name: String, desc: Option<String>, addresses: Vec<Address>) -> Device {
+    fn new(
+        name: String,
+        desc: Option<String>,
+        addresses: Vec<Address>,
+        flags: DeviceFlags,
+    ) -> Device {
         Device {
             name,
             desc,
             addresses,
+            flags,
         }
     }
 
@@ -236,7 +339,7 @@ impl Device {
         Capture::from_device(self)?.open()
     }
 
-    /// Returns the default Device suitable for captures according to pcap_lookupdev,
+    /// Returns the default Device suitable for captures according to pcap_findalldevs,
     /// or an error from pcap. Note that there may be no suitable devices.
     pub fn lookup() -> Result<Option<Device>, Error> {
         unsafe {
@@ -285,7 +388,7 @@ impl Device {
 
 impl From<&str> for Device {
     fn from(name: &str) -> Self {
-        Device::new(name.into(), None, Vec::new())
+        Device::new(name.into(), None, Vec::new(), DeviceFlags::empty())
     }
 }
 
@@ -297,6 +400,7 @@ impl TryFrom<&raw::pcap_if_t> for Device {
             unsafe { cstr_to_string(dev.name)?.ok_or(InvalidString)? },
             unsafe { cstr_to_string(dev.description)? },
             unsafe { Address::new_vec(dev.addresses) },
+            DeviceFlags::from(dev.flags),
         ))
     }
 }