+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add `PacketStream::capture_mut` to still be able to inject packets when using `PacketStream`
 - `Capture::iter()` that return an iterator that use a codec like `Capture::stream()`
  - Add `Packet<Dead>::dead_with_precision` to enable creating a pcap with nanosecond precision.
+ - Add `flags` field to `Device`.
 
 ### Changed